backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
google-genai>=1.45.0
python-multipart==0.0.12
python-dotenv==1.0.1
Pillow==11.0.0
psycopg2-binary==2.9.11
numpy>=1.24.0
onnxruntime>=1.18.0
rembg>=2.0.0

